+++ test/runtests.jl
@@ -4,6 +4,7 @@ using LinearAlgebra
 using Logging
 using Makie
 using OrdinaryDiffEq
+using OrdinaryDiffEqAdamsBashforthMoulton
 using StaticArrays
 using SatelliteAnalysis